% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTSEM-package.R
\docType{package}
\name{QTSEM-package}
\alias{QTSEM-package}
\title{QTSEM: A Package about Qingtai-Tibet SEM Modelling on Travel Behavior}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.

This package is an exercise in package creation using
R studio. The package includes a sample function and
a sample dataset with their respective documentation.
}
\references{
\url{https://github.com/paezha/packr}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Horan517/QTSEM}
  \item Report bugs at \url{https://github.com/Horan517/QTSEM/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/Horan517/QTSEM}
  \item Report bugs at \url{https://github.com/Horan517/QTSEM/issues}
}

}
\author{
Haoran Xu, School of Earth, Environment and Society, McMaster University \email{xu976@mcmaster.ca}
}
\keyword{internal}
