% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTSEM-package.R
\docType{package}
\name{QTSEM-package}
\alias{QTSEM}
\alias{QTSEM-package}
\title{QTSEM: A Package about Conducting Structural Equation Modeling Analysis for Travel Behavior on Qinghai-Tibet Plateau}
\description{
Structural equation modeling (SEM) is a statistical technique that combines factor analysis and multiple regression to evaluate and test complex relationships between observed and latent variables within a hypothesized model. This package utilizes this method to analyze a travel behavior and satisfaction-related questionnaire dataset that was collected during July to August on Qinghai-Tibet Plateau.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Horan517/QTSEM}
  \item Report bugs at \url{https://github.com/Horan517/QTSEM/issues}
}

}
\keyword{internal}
